+++ src/lighteval/models/abstract_model.py
@@ -27,7 +27,6 @@ from typing import Optional, Union
 import torch
 from transformers import BatchEncoding
 
-from lighteval.models.model_config import EnvConfig
 from lighteval.models.model_output import (
     GenerateMultiTurnReturn,
     GenerateReturn,
@@ -41,6 +40,7 @@ from lighteval.tasks.requests import (
     LoglikelihoodRollingRequest,
     LoglikelihoodSingleTokenRequest,
 )
+from lighteval.utils import EnvConfig
 
 
 TokenSequence = Union[list[int], torch.LongTensor, torch.Tensor, BatchEncoding]
